Allow saving of profile data in session variables
Profile data (favourite colour, fruit) can be saved on profile tab if valid user logged in. The fields are populated with current choice and remain so even if user leaves page (tab) & revisits. If the Profile page is visited by user not validly logged in then the page tells them so & they are not given chance to enter profile data.